Much more efficient queries
Previously, we would tediously scrape for every job, and for every
job look up details for every build, making hundreds or thousands of
HTTP requests to get a single full scan.

This new approach uses the `tree` parameters in the jenkins API
which lets us get almost all the data we need in a single query.  We
still need to make multiple queries to recurse into "folders" of
jobs.